--- a/tsdb/index/tsi1/index_file.go
+++ b/tsdb/index/tsi1/index_file.go
@@ -64,8 +64,8 @@ type IndexFile struct {
 	level int
 	id    int
 
+	mu sync.RWMutex
 	// Compaction tracking.
-	mu         sync.RWMutex
 	compacting bool
 
 	// Path to data file.
@@ -244,7 +244,7 @@ func (f *IndexFile) SeriesIDSet() (*tsdb.SeriesIDSet, error) {
 
 func (f *IndexFile) TombstoneSeriesIDSet() (*tsdb.SeriesIDSet, error) {
 	ss := tsdb.NewSeriesIDSet()
-	if err := ss.UnmarshalBinary(f.tombstoneSeriesIDSetData); err != nil {
+	if err := ss.UnmarshalBinaryUnsafe(f.tombstoneSeriesIDSetData); err != nil {
 		return nil, err
 	}
 	return ss, nil