@hercules-influx 2018-08-09
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/654492739/src/github.com/influxdata/influxdb/tsdb/index/tsi1/log_file_test.go:75:7: this value of err is never used (SA4006)
/tmp/654492739/src/github.com/influxdata/influxdb/tsdb/index/tsi1/index.go:1075:6: type ssElement is unused (U1000)
```