@zeebo 2018-08-09
This is really cool. It looks like you took the path of attempting to keep the cache updated as changes happen. Do you think it's worthwhile to instead assume a more steady state and carefully evict some entries if we detect they could have changed? Then we don't have to worry about keeping the data in sync with the source of truth.