@e-dard 2018-08-09
@zeebo I have updated. My concern with evicting on change is that for our internal projects where measurement would be an entire database of data, any new series would evict the entire cache, which would defeat its purpose in the first place.