@hercules-influx 2018-08-09
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/273743947/src/github.com/influxdata/influxdb/tsdb/index/tsi1/log_file_test.go:75:7: this value of err is never used (SA4006)
```